Update file structure to current template
