+++ labs/HelloWorld/.ask/config
@@ -1,0 +1,19 @@
+{
+  "deploy_settings": {
+    "default": {
+      "skill_id": "",
+      "was_cloned": false,
+      "merge": {
+        "skillManifest": {
+          "apis": {
+            "custom": {
+              "endpoint": {
+                "uri": "ask-custom-helloworld-default"
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
