+++ labs/HelloWorld/lambda/custom/index.js
@@ -1,0 +1,44 @@
+const Alexa = require('alexa-sdk');
+
+exports.handler = function(event, context, callback) {
+    let alexa = Alexa.handler(event, context);
+
+    // alexa.appId = 'amzn1.echo-sdk-ams.app.1234';
+
+    ///alexa.dynamoDBTableName = 'YourTableName'; // creates new table for session.attributes
+
+    alexa.registerHandlers(handlers);
+    alexa.execute();
+};
+
+const HELLO_MESSAGE = 'Hello World from Alexa'
+const HELP_MESSAGE = 'You can say hello, or, you can say exit... What can I help you with?';
+const HELP_REPROMPT = 'What can I help you with?';
+const STOP_MESSAGE = 'Goodbye!';
+
+const handlers = {
+    'LaunchRequest': function () {
+        this.emit('MyIntent');
+    },
+    'MyIntent': function () {
+        const speechOutput = HELLO_MESSAGE;
+
+        this.response.speak(speechOutput);
+        this.emit(':responseReady');
+    },
+    'AMAZON.HelpIntent': function () {
+        const speechOutput = HELP_MESSAGE;
+        const reprompt = HELP_REPROMPT;
+
+        this.response.speak(speechOutput).listen(reprompt);
+        this.emit(':responseReady');
+    },
+    'AMAZON.CancelIntent': function () {
+        this.response.speak(STOP_MESSAGE);
+        this.emit(':responseReady');
+    },
+    'AMAZON.StopIntent': function () {
+        this.response.speak(STOP_MESSAGE);
+        this.emit(':responseReady');
+    },
+};
