+++ labs/HelloWorld/lambda/tests/test.js
@@ -1,10 +1,10 @@
 // This is a Javascript test harness that simulates the execution of Lambda function code
 // From the command prompt, type "node test.js"
 
-var MyLambdaFunction = require('../src/index.js'); // assumes single Lambda function with exports.handler
+const MyLambdaFunction = require('../src/index.js'); // assumes single Lambda function with exports.handler
 
  // this is the Lambda request data, generated by the Alexa service.  Replace with your actual Alexa request.
-var event =
+const event =
     {
         "session": {
             "sessionId": "SessionId.f9e6dcbb-b7da-4b47-905c.etc.etc",
@@ -25,7 +25,7 @@ var event =
         "version": "1.0"
     };
 
-var context = {
+let context = {
     'succeed': function (data) {
         console.log(JSON.stringify(data, null,'\t') );
