+++ labs/HelloWorld/set-up-aws.md
@@ -1,0 +1,15 @@
+# Create an AWS Account
+ 
+  **Note: If you already have an AWS account, you can skip this section.  Just sign in to your [console](http://aws.amazon.com).**
+
+1.  Open [aws.amazon.com](aws.amazon.com) and then choose **‘Create an AWS Account’**
+
+    <img src="https://m.media-amazon.com/images/G/01/mobile-apps/dex/alexa/alexa-skills-kit/tutorials/general/buttons/create-an-aws-account-button._TTH_.png" />
+
+    1. Follow the online instructions. Do not worry about the IAM role, we will do that later.
+    2. You will need a Valid Credit Card to set up your account (note the AWS Free Tier will suffice however. [You can find out more about the free tier here](https://aws.amazon.com/free/?sc_ichannel=ha&amp;sc_ipage=signin&amp;sc_iplace=body_link_text&amp;sc_icampaigntype=free_tier&amp;sc_icampaign=ha_en_free_tier_signin_2014_03).)
+    3. Part of the sign-up procedure involves receiving a phone call and entering a PIN using the phone keypad.
+    
+2.  Sign in to the AWS Console
+
+3.  It can sometimes take a couple minutes for your new AWS account to go live. You will receive an e-mail when your account is active.
