--- labs/HelloWorld/speechAssets/IntentSchema.json
@@ -1,17 +1,0 @@
-{
-  "intents": [
-    {
-      "intent": "MyIntent",  "slots":[]
-    },
-
-    {
-      "intent": "AMAZON.HelpIntent"
-    },
-    {
-      "intent": "AMAZON.StopIntent"
-    },
-    {
-      "intent": "AMAZON.CancelIntent"
-    }
-  ]
-}
