--- labs/HelloWorld/speechAssets/SampleUtterances.txt
@@ -1,1 +1,0 @@
-MyIntent hello
