--- labs/HelloWorld/src/index.js
@@ -1,23 +1,0 @@
-const Alexa = require('alexa-sdk');
-
-exports.handler = function(event, context, callback) {
-    var alexa = Alexa.handler(event, context);
-
-    // alexa.appId = 'amzn1.echo-sdk-ams.app.1234';
-
-    ///alexa.dynamoDBTableName = 'YourTableName'; // creates new table for session.attributes
-
-    alexa.registerHandlers(handlers);
-    alexa.execute();
-};
-
-const handlers = {
-    'LaunchRequest': function () {
-        this.emit('MyIntent');
-    },
-
-    'MyIntent': function () {
-        this.response.speak('Hello World from Alexa');
-        this.emit(':responseReady')
-    }
-};
